Make CWS multi-contest
