+++ cms/server/contest/handlers/__init__.py
@@ -9,7 +9,7 @@
 # Copyright © 2013 Bernard Blackham <bernard@largestprime.net>
 # Copyright © 2014 Artem Iglikov <artem.iglikov@gmail.com>
 # Copyright © 2014 Fabian Gundlach <320pointsguy@gmail.com>
-# Copyright © 2015 William Di Luigi <williamdiluigi@gmail.com>
+# Copyright © 2015-2016 William Di Luigi <williamdiluigi@gmail.com>
 #
 # This program is free software: you can redistribute it and/or modify
 # it under the terms of the GNU Affero General Public License as
@@ -33,7 +33,6 @@ from cms import config
 from .base import \
     StaticFileGzHandler
 from .main import \
-    MainHandler, \
     LoginHandler, \
     LogoutHandler, \
     StartHandler, \
@@ -67,7 +66,6 @@ HANDLERS = [
 
     # Main
 
-    (r"/", MainHandler),
     (r"/login", LoginHandler),
     (r"/logout", LogoutHandler),
     (r"/start", StartHandler),
