+++ cms/server/contest/handlers/communication.py
@@ -9,7 +9,7 @@
 # Copyright © 2013 Bernard Blackham <bernard@largestprime.net>
 # Copyright © 2014 Artem Iglikov <artem.iglikov@gmail.com>
 # Copyright © 2014 Fabian Gundlach <320pointsguy@gmail.com>
-# Copyright © 2015 William Di Luigi <williamdiluigi@gmail.com>
+# Copyright © 2015-2016 William Di Luigi <williamdiluigi@gmail.com>
 #
 # This program is free software: you can redistribute it and/or modify
 # it under the terms of the GNU Affero General Public License as
@@ -33,39 +33,46 @@ from __future__ import print_function
 from __future__ import unicode_literals
 
 import logging
+import os
 
 import tornado.web
 
 from cms.db import Question
+from cms.server import multi_contest
 
-from .base import BaseHandler, NOTIFICATION_ERROR, NOTIFICATION_SUCCESS
+from .contest import ContestHandler, NOTIFICATION_ERROR, NOTIFICATION_SUCCESS
 
 
 logger = logging.getLogger(__name__)
 
 
-class CommunicationHandler(BaseHandler):
+class CommunicationHandler(ContestHandler):
     """Displays the private conversations between the logged in user
     and the contest managers..
 
     """
     @tornado.web.authenticated
-    def get(self):
-        self.set_secure_cookie("unread_count", "0")
+    @multi_contest
+    def get(self, contest_name):
+        self.set_secure_cookie(self.contest.name + "_unread_count", "0")
         self.render("communication.html", **self.r_params)
 
 
-class QuestionHandler(BaseHandler):
+class QuestionHandler(ContestHandler):
     """Called when the user submits a question.
 
     """
     @tornado.web.authenticated
-    def post(self):
+    @multi_contest
+    def post(self, contest_name):
         participation = self.current_user
 
         # User can post only if we want.
         if not self.contest.allow_questions:
             raise tornado.web.HTTPError(404)
+
+        fallback_page = os.path.join(self.r_params["real_contest_root"],
+                                     "communication")
 
         subject_length = len(self.get_argument("question_subject", ""))
         text_length = len(self.get_argument("question_text", ""))
@@ -79,7 +86,7 @@ class QuestionHandler(BaseHandler):
                 self._("Question too big!"),
                 self._("You have reached the question length limit."),
                 NOTIFICATION_ERROR)
-            self.redirect("/communication")
+            self.redirect(fallback_page)
             return
 
         question = Question(self.timestamp,
@@ -101,4 +108,4 @@ class QuestionHandler(BaseHandler):
                    "notified when it is answered."),
             NOTIFICATION_SUCCESS)
 
-        self.redirect("/communication")
+        self.redirect(fallback_page)
