+++ cms/server/contest/templates/contest_list.html
@@ -1,0 +1,16 @@
+{% extends base.html %}
+{% block title %}
+    Choose a contest
+{% end %}
+{% block body %}
+    <div class="hero-unit contest-list">
+        <h3>Choose a contest</h3>
+        <ul class="nav nav-list">
+    {% for c_name, c_iter in contest_list.iteritems() %}
+            <li>
+                <a {% if c_iter.stop < datetime.datetime.now() %} style="color: #adadad" {% end %} href="{{ url_root }}/{{ c_iter.name }}">{{ c_iter.description }}</a>
+            </li>
+    {% end %}
+        </ul>
+    </div>
+{% end %}
