+++ cms/server/contest/templates/documentation.html
@@ -1,4 +1,4 @@
-{% extends base.html %}
+{% extends contest.html %}
 {% block core %}
 <div class="span9">
 
@@ -8,7 +8,7 @@
 
 <h2>{{ _("Programming languages and libraries") }}</h2>
 
-<a href="{{ url_root }}/stl/index.html">{{ _("Standard Template Library") }}</a>
+<a href="{{ contest_root }}/stl/index.html">{{ _("Standard Template Library") }}</a>
 
 <h2>{{ _("Submission details for compilation") }}</h2>
