+++ cms/server/contest/templates/printing.html
@@ -1,4 +1,4 @@
-{% extends base.html %}
+{% extends contest.html %}
 
 {% block core %}
 
@@ -20,7 +20,7 @@
     {% end %}
 
 <div id="print" class="row">
-    <form class="form-horizontal" enctype="multipart/form-data" action="{{ url_root }}/printing" method="POST">
+    <form class="form-horizontal" enctype="multipart/form-data" action="{{ contest_root }}/printing" method="POST">
         <fieldset>
             <div class="control-group">
                 <label class="control-label" for="input">{% if pdf_printing_allowed %}{{ _("File (text or PDF)") }}{% else %}{{ _("File (text)") }}{% end %}</label>
