+++ cms/server/contest/templates/submission_row.html
@@ -92,7 +92,7 @@
             {% if s.language is not None %}
                 {% set filename = re.sub("\.%l$", ".%s" % s.language, filename) %}
             {% end %}
-        <a class="btn" href="{{ url_root }}/tasks/{{ encode_for_url(task.name) }}/submissions/{{ s_idx }}/files/{{ encode_for_url(filename) }}">
+        <a class="btn" href="{{ contest_root }}/tasks/{{ encode_for_url(task.name) }}/submissions/{{ s_idx }}/files/{{ encode_for_url(filename) }}">
             {{ _("Download") }}
         </a>
         {% else %}
@@ -110,7 +110,7 @@
                     {% end %}
                 {% end %}
                 <li>
-                    <a href="{{ url_root }}/tasks/{{ encode_for_url(task.name) }}/submissions/{{ s_idx }}/files/{{ encode_for_url(filename) }}">
+                    <a href="{{ contest_root }}/tasks/{{ encode_for_url(task.name) }}/submissions/{{ s_idx }}/files/{{ encode_for_url(filename) }}">
                         {{ filename }}
                     </a>
                 </li>
@@ -127,7 +127,7 @@
         {% else %}
             {% comment TODO If the expiration or generation time are greater than valid_phase_end they "don't exist". It's useless to show "Wait..." in those cases. %}
             {% if can_play_token and not need_to_wait %}
-        <form action="{{ url_root }}/tasks/{{ encode_for_url(task.name) }}/submissions/{{ s_idx }}/token" method="POST">
+        <form action="{{ contest_root }}/tasks/{{ encode_for_url(task.name) }}/submissions/{{ s_idx }}/token" method="POST">
             <button type="submit" class="btn btn-warning">{{ _("Play!") }}</button>
         </form>
             {% elif (can_play_token and need_to_wait) or (not can_play_token and tokens_info[1] is not None) %}
