+++ cms/server/contest/templates/user_test_row.html
@@ -50,7 +50,7 @@
     </td>
 {% end %}
     <td class="input">
-        <a class="btn" href="{{ url_root }}/tasks/{{ encode_for_url(task.name) }}/tests/{{ t_idx }}/input">
+        <a class="btn" href="{{ contest_root }}/tasks/{{ encode_for_url(task.name) }}/tests/{{ t_idx }}/input">
             {{ _("Download") }}
         </a>
     </td>
@@ -65,7 +65,7 @@
             {{ _("N/A") }}
         </a>
         {% else %}
-        <a class="btn btn-primary" href="{{ url_root }}/tasks/{{ encode_for_url(task.name) }}/tests/{{ t_idx }}/output">
+        <a class="btn btn-primary" href="{{ contest_root }}/tasks/{{ encode_for_url(task.name) }}/tests/{{ t_idx }}/output">
             {{ _("Download") }}
         </a>
         {% end %}
@@ -82,7 +82,7 @@
         </a>
         {% elif len(files) == 1 %}
             {% set filename = t.files.keys()[0] %}
-        <a class="btn" href="{{ url_root }}/tasks/{{ encode_for_url(task.name) }}/tests/{{ t_idx }}/files/{{ encode_for_url(filename) }}">
+        <a class="btn" href="{{ contest_root }}/tasks/{{ encode_for_url(task.name) }}/tests/{{ t_idx }}/files/{{ encode_for_url(filename) }}">
             {{ _("Download") }}
         </a>
         {% else %}
@@ -99,7 +99,7 @@
                     {% set real_filename = filename %}
                 {% end %}
                 <li>
-                    <a href="{{ url_root }}/tasks/{{ encode_for_url(task.name) }}/tests/{{ t_idx }}/files/{{ encode_for_url(filename) }}">
+                    <a href="{{ contest_root }}/tasks/{{ encode_for_url(task.name) }}/tests/{{ t_idx }}/files/{{ encode_for_url(filename) }}">
                         {{ real_filename }}
                     </a>
                 </li>
